requirements.txt
absl-py==0.1.10
atari-py==0.1.1
bleach==1.5.0
certifi==2018.1.18
chardet==3.0.4
future==0.16.0
gym==0.7.3
html5lib==0.9999999
idna==2.6
Keras==2.1.4
Markdown==2.6.11
numpy==1.14.1
Pillow==5.0.0
protobuf==3.5.1
pyglet==1.3.1
PyOpenGL==3.1.0
PyYAML==3.12
requests==2.18.4
scipy==1.0.0
six==1.11.0
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
urllib3==1.22
Werkzeug==0.14.1
